test_1/scrapy/1_urls/requirements.txt
scrapy
scrapy-splash
scrapy-selenium
cookiecutter
selenium
scrapy_cloudflare_middleware
Pillow
scrapy-rotating-proxies
pandas
python-dotenv
psycopg2
scrapyrt